+++ __tests__/Validation.test.js
@@ -1,5 +1,5 @@
-import Validation from '../src/utils/Validation';
-import { ERROR_MESSAGE } from '../src/constants';
+import Validation from '../src/Validation';
+import { ERROR_MESSAGE } from '../src/constants/message';
 
 describe('사용자 입력 유효성 검사', () => {
   test.each([[['pobi', 'jun', '']], [['pobipobi', 'jun', 'cron']]])(
