+++ src/constants/number.js
@@ -1,0 +1,10 @@
+const NUMBERS = Object.freeze({
+  CAR_NAME_MINIMUM_LENGTH: 1,
+  CAR_NAME_MAXIMUM_LENGTH: 5,
+  TRY_COUNT_MINIMUM_COUNT: 0,
+  DEFAULT_CAR_DISTANCE: 0,
+  MINIMUM_MOVEMENT_THRESHOLD: 4,
+  MOVE_INCREMENT: 1,
+});
+
+export { NUMBERS };
